third_party/WebKit/LayoutTests/http/tests/loading/empty-content-disposition-type.html

<!doctype html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test(t => {
  const url = 'resources/empty-content-disposition-type.php';
  const iframe = document.createElement('iframe');
  iframe.addEventListener('load', t.step_func(() => {
    assert_equals(iframe.contentDocument.body.textContent, 'hello, world\n');
    t.done();
  }));
  iframe.src = url;
  document.body.appendChild(iframe);
}, 'load a page with an invalid content-disposition-type');
</script>
</body>
